refactor: expose root and fix importsUpdate exports make the package root point to the builtmodule (dist/index.mjs) so consumers and code can import fromthe package root instead a path Remove the redundant"./zero-id" export entry.
Adjust internal and tests to import from the package rootrather than "./zero-id" or "../src/zero-id This simplifies moduleresolution and keeps import paths consistent across source and tests.

Also update internal demo import in generate.ts to useroot index export.
