+++ packages/zero-id/tests/genetate-zero-id.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "vitest";
-import { generateZeroId } from "../src/zero-id";
+import { generateZeroId } from "../src";
 
 const ZERO_ID_PATTERN = /^zero-\d{8}-([0-9a-f]{6}-){3}[0-9a-f]{6}$/;
